Moved exceptions to the correct package
