+++ de.persosim.simulator/src/de/persosim/simulator/cardobjects/ShortFileIdentifier.java
@@ -1,5 +1,7 @@
 package de.persosim.simulator.cardobjects;
 
+import de.persosim.simulator.exception.FileIdentifierIncorrectValueException;
+
 
 /**
  * Implementation of an ISO7816-4 compliant short identifier.
