+++ de.persosim.simulator/src/de/persosim/simulator/exception/ObjectNotModifiedException.java
@@ -1,4 +1,4 @@
-package de.persosim.simulator.cardobjects;
+package de.persosim.simulator.exception;
 
 public class ObjectNotModifiedException extends RuntimeException {
