+++ de.persosim.simulator/src/de/persosim/simulator/protocols/file/AbstractFileProtocol.java
@@ -10,10 +10,10 @@ import de.persosim.simulator.cardobjects.CardObject;
 import de.persosim.simulator.cardobjects.DedicatedFileIdentifier;
 import de.persosim.simulator.cardobjects.ElementaryFile;
 import de.persosim.simulator.cardobjects.FileIdentifier;
-import de.persosim.simulator.cardobjects.FileIdentifierIncorrectValueException;
 import de.persosim.simulator.cardobjects.Scope;
 import de.persosim.simulator.cardobjects.ShortFileIdentifier;
 import de.persosim.simulator.exception.AccessDeniedException;
+import de.persosim.simulator.exception.FileIdentifierIncorrectValueException;
 import de.persosim.simulator.exception.FileToShortException;
 import de.persosim.simulator.exception.TagNotFoundException;
 import de.persosim.simulator.platform.CardStateAccessor;
